Refactoring code, improve HTTP status and error handling
